src/requirements.txt
aiohttp==3.8.1
aioredis==1.3.1
aiosignal==1.2.0
aniso8601==7.0.0
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.4.0
autobahn==22.4.2
Automat==20.2.0
bidict==0.22.0
certifi==2022.5.18.1
cffi==1.15.0
channels==3.0.4
channels-redis==3.4.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
constantly==15.1.0
cryptography==37.0.2
daphne==3.0.2
Django==4.0.4
django-channels==0.7.0
django-dotenv==1.4.2
frozenlist==1.3.0
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
h11==0.13.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
msgpack==1.0.3
multidict==6.0.2
oauthlib==3.2.0
Pillow==9.1.1
promise==2.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyee==9.0.4
pyOpenSSL==22.0.0
python-engineio==4.3.2
python-socketio==5.6.0
requests==2.27.1
requests-oauthlib==1.3.1
Rx==1.6.1
service-identity==21.1.0
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
Twisted==22.4.0
twisted-iocpsupport==1.0.2
txaio==22.2.1
typing_extensions==4.2.0
tzdata==2022.1
urllib3==1.26.9
uvicorn==0.17.6
websockets==10.3
yarl==1.7.2
zope.interface==5.4.0
